requirements.txt
getgauge==0.3.17
selenium==4.16.0
webdriver-manager==4.0.1
requests==2.31.0
Appium-Python-Client==3.1.1
